// src/screens/shared/CustomerInformationForm/types.ts
export interface CustomerInformationFormValues {
  addressAddon: string;
  city: string;
  commercialRegisterNumber: string;
  commercialRegisterNumberRequired?: boolean;
  companyId: string;
  companyName: string;
  country: string;
  legalForm: string;
  postcode: string;
  registrationAuthority: string;
  streetAndNumber: string;
  vatId: string;
}